Add agent check-in interval to agent install tests
Omitting the agent check-in interval when patching the
`fleet-controller` config map would now lead to errors when setting up
agents with a check-in interval bearing the default value for a
duration, ie 0s.
That interval is now set with a hard-coded value, which is of no
importance for such tests, for the sake of not being zero.