[WIP] Detect overlaps between bundles of a same GitRepo
This commit enables `fleet apply` to detect any possible resource
overlaps, by kind and name, between:
* obsolete bundles, which will be deleted because they no longer match
  any path in the GitRepo
* new bundles coming from the GitRepo, to be created

When such overlaps are detected, the bundle to be written has a newly
populated `Overwrites` field, which is propagated to bundle deployments
when targeting that bundle, specifying which resources (name, kind and
namespace) this bundle deployment will overwrite from other bundles.

This enables the agent to detect if a bundle deployment's
`ModifiedStatus` features missing resources which may have been deleted
when deleting a previous, obsolete bundle, for the bundle deployment to
be redeployed. This last part does not quite work yet:
* new deployments are triggered more than once
* they do not result in updated, non-modified bundle deployment
  statuses.

As a side note, it is unclear whether namespaces can even be reliably
populated in bundle deployment overwrites, as a bundle resource's actual
target namespace may depend on multiple factors.